Property Function Implemented (draft)

+++ src/edu/upenn/cit594/processor/PropertyFunction.java
@@ -1,7 +1,9 @@
 package edu.upenn.cit594.processor;
+
 import java.util.List;
+
 import edu.upenn.cit594.util.PropertyData;
 
 public interface PropertyFunction {
-    int compute(List<PropertyData> properties);
+    double compute(List<PropertyData> properties);
 }
